feat: move client code to seperate module
